% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data2max.R
\name{data2max}
\alias{data2max}
\title{data2max
Maximized the data.}
\usage{
data2max(data, mth, zmth, xbest, qlow, qup)
}
\arguments{
\item{data}{The Data that needs to be maximized.}

\item{mth}{A vector describing the type of each indicator.}

\item{zmth}{A vector describing the method from negative to positive, including DS: reciprocal method, it is not recommended to use when the data has 0, CZ: using the maximum subtraction method, other types of indicators are marked as NA.}

\item{xbest}{An optimal value vector of intermediate numerical values,other types of indicators are marked as NA.}

\item{qlow}{An interval lower bound vector with interval numeric value forward,other types of indicators are marked as NA.}

\item{qup}{An interval upper bound vector of the interval value maximize,other types of indicators are marked as NA.}
}
\value{
A maximized data set
}
\description{
data2max
Maximized the data.
}
\examples{
\donttest{
sampleData<-sampleData
mth<-c("ZH","ZJ","QJ","ZH","FU","ZH","QJ","FU","FU","ZH")
zmth<-c(NA,NA,NA,NA,"CZ",NA,NA,"DS","CZ",NA)
xbest<-c(NA,4,NA,NA,NA,NA,NA,NA,NA,NA)
qup<-c(NA,NA,5,NA,NA,NA,5,NA,NA,NA)
qlow<-c(NA,NA,3,NA,NA,NA,3,NA,NA,NA)
data2max(sampleData,mth,zmth,xbest,qlow,qup)
}
}
