% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataNormal.R
\name{DataNormal}
\alias{DataNormal}
\title{DataNormal
Normalize the maximized data set.}
\usage{
DataNormal(data)
}
\arguments{
\item{data}{A maximized data set}
}
\value{
A Normalized data set
}
\description{
DataNormal
Normalize the maximized data set.
}
\examples{
\dontrun{
DataNormal(data2max(sampleData))
}
}
