% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataNormalpre.R
\name{DataNormalpre}
\alias{DataNormalpre}
\title{DataNormalpre
Prepare the normalized data set.}
\usage{
DataNormalpre(data, i)
}
\arguments{
\item{data}{A maximized data set}

\item{i}{Index column}
}
\value{
Index column normalized
}
\description{
DataNormalpre
Prepare the normalized data set.
}
