% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ejcaul.R
\name{Ejcaul}
\alias{Ejcaul}
\title{Ejcaul
Calculate Ej value.}
\usage{
Ejcaul(data, i)
}
\arguments{
\item{data}{A normalized data set.}

\item{i}{Index column.}
}
\value{
Index column weight Ej
}
\description{
Ejcaul
Calculate Ej value.
}
