% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Errmax.R
\name{Errmax}
\alias{Errmax}
\title{Errmax
Function on error.}
\usage{
Errmax(x, i)
}
\arguments{
\item{x}{The Data that needs to be maximized.}

\item{i}{Index column.}
}
\value{
Index column maximized.
}
\description{
Errmax
Function on error.
}
