% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max2max.R
\name{max2max}
\alias{max2max}
\title{max2max
positive Data conversion.}
\usage{
max2max(x, i)
}
\arguments{
\item{x}{The Data that needs to be maximized.}

\item{i}{Index column.}
}
\value{
Index column maximized
}
\description{
max2max
positive Data conversion.
}
