% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min2maxC.R
\name{min2maxC}
\alias{min2maxC}
\title{min2maxC
negative Data conversion use Maximum subtraction method}
\usage{
min2maxC(x, i)
}
\arguments{
\item{x}{The Data that needs to be maximized.}

\item{i}{Index column.}
}
\value{
Index column maximized
}
\description{
min2maxC
negative Data conversion use Maximum subtraction method
}
