% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min2maxD.R
\name{min2maxD}
\alias{min2maxD}
\title{min2maxD
negative Data conversion use reciprocal method}
\usage{
min2maxD(x, i)
}
\arguments{
\item{x}{The Data that needs to be maximized.}

\item{i}{Index column.}
}
\value{
Index column maximized
}
\description{
min2maxD
negative Data conversion use reciprocal method
}
