% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qj2max.R
\name{qj2max}
\alias{qj2max}
\title{qj2max
Interval Data conversion}
\usage{
qj2max(x, qlow, qup, i)
}
\arguments{
\item{x}{The Data that needs to be maximized.}

\item{qlow}{An interval lower bound vector with interval numeric value forward,other types of indicators are marked as NA.}

\item{qup}{An interval upper bound vector of the interval value maximize,other types of indicators are marked as NA.}

\item{i}{Index column.}
}
\value{
Index column maximized
}
\description{
qj2max
Interval Data conversion
}
