% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samlpeData.R
\docType{data}
\name{sampleData}
\alias{sampleData}
\title{A example of multiple-criteria decision making data.}
\format{
A data fram with 21 rows and 11 variables.
}
\source{
Randomly generated.
}
\usage{
sampleData
}
\description{
A example of multiple-criteria decision making data.
}
\keyword{datasets}
