% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zj2max.R
\name{zj2max}
\alias{zj2max}
\title{zj2max
Intermediate Data conversion}
\usage{
zj2max(x, xbest, i)
}
\arguments{
\item{x}{The Data that needs to be maximized.}

\item{xbest}{An optimal value vector of intermediate numerical values,other types of indicators are marked as NA.}

\item{i}{Index column.}
}
\value{
Index column maximized
}
\description{
zj2max
Intermediate Data conversion
}
