% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataNormal.R
\name{DataNormal}
\alias{DataNormal}
\title{DataNormal
Normalize the maximized data set.}
\usage{
DataNormal(data)
}
\arguments{
\item{data}{A maximized data set}
}
\value{
A Normalized data set
}
\description{
DataNormal
Normalize the maximized data set.
}
\examples{
\dontrun{
sampleData<-sampleData
mth<-c("ZH","ZJ","QJ","ZH","FU","ZH","QJ","FU","FU","ZH")
zmth<-c(NA,NA,NA,NA,"CZ",NA,NA,"DS","CZ",NA)
xbest<-c(NA,4,NA,NA,NA,NA,NA,NA,NA,NA)
qup<-c(NA,NA,5,NA,NA,NA,5,NA,NA,NA)
qlow<-c(NA,NA,3,NA,NA,NA,3,NA,NA,NA)
maxdata<-data2max(sampleData,mth,zmth,xbest,qlow,qup)
DataNormal(maxdata)
}
}
