% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetWeight.R
\name{GetWeight}
\alias{GetWeight}
\title{GetWeight
Calculate weights on standardized data us the entropy method.}
\usage{
GetWeight(data)
}
\arguments{
\item{data}{A Normalized data set}
}
\value{
WeightEij is Ej value of Index,weight is weight value of Index
}
\description{
GetWeight
Calculate weights on standardized data us the entropy method.
}
\examples{
\donttest{
sampleData<-sampleData
mth<-c("ZH","ZJ","QJ","ZH","FU","ZH","QJ","FU","FU","ZH")
zmth<-c(NA,NA,NA,NA,"CZ",NA,NA,"DS","CZ",NA)
xbest<-c(NA,4,NA,NA,NA,NA,NA,NA,NA,NA)
qup<-c(NA,NA,5,NA,NA,NA,5,NA,NA,NA)
qlow<-c(NA,NA,3,NA,NA,NA,3,NA,NA,NA)
weightdata<-data2max(sampleData,mth,zmth,xbest,qlow,qup)
GetWeight(weightdata)
}
}
