% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetWeight.R
\name{GetWeight}
\alias{GetWeight}
\title{GetWeight
Calculate weights on standardized data us the entropy method.}
\usage{
GetWeight(data)
}
\arguments{
\item{data}{A Normalized data set}
}
\value{
WeightEij is Ej value of Index,weight is weight value of Index
}
\description{
GetWeight
Calculate weights on standardized data us the entropy method.
}
\examples{
\dontrun{
i=10
GetWeight(sampleData)
}
}
