% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WtTopsisN.R
\name{WtTopsisN}
\alias{WtTopsisN}
\title{WtTopsis
Weighted TOPSIS method for multiple-criteria decision making (MCDM) us Normalized data.}
\usage{
WtTopsisN(data)
}
\arguments{
\item{data}{Normalized data.}
}
\value{
Data set containing D+,D- and C values
}
\description{
WtTopsis
Weighted TOPSIS method for multiple-criteria decision making (MCDM) us Normalized data.
}
\examples{
\dontrun{
WtTopsisN<-(sampleData)
}
}
